// apps/admin/src/api/generated/model/adminOrderFulfillmentStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Admin API
 * Back-office API for staff: catalog, orders, fulfillment, payments and users.
 * OpenAPI spec version: 0.1.0
 */

export type AdminOrderFulfillmentStatus = typeof AdminOrderFulfillmentStatus[keyof typeof AdminOrderFulfillmentStatus];


export const AdminOrderFulfillmentStatus = {
  unfulfilled: 'unfulfilled',
  fulfilled: 'fulfilled',
  shipped: 'shipped',
  delivered: 'delivered',
} as const;