// apps/admin/src/api/generated/model/adminOrderStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Admin API
 * Back-office API for staff: catalog, orders, fulfillment, payments and users.
 * OpenAPI spec version: 0.1.0
 */

export type AdminOrderStatus = typeof AdminOrderStatus[keyof typeof AdminOrderStatus];


export const AdminOrderStatus = {
  pending: 'pending',
  completed: 'completed',
  canceled: 'canceled',
  archived: 'archived',
} as const;